chore: fix all clippy warnings across workspace
- Replace manual Default impls with derive (knit-core, knit-noise)
- Use iterator patterns instead of index-based loops (knit-noise, knit-gen, knit-learn)
- Replace clone-to-slice with std::slice::from_ref (knit-gen engine)
- Use Range::contains / div_ceil / is_multiple_of (knit-plan, knit-learn)
- Replace map_err with inspect_err, map_or with simplified form (knit-learn, integration-tests)
- Fix digit grouping and approximate-PI lint in test values (knit-bind, knit-core)
- Use returning-let-binding simplification (knit-cli)
- Remove unnecessary casts (knit-bind json)

+++ crates/knit-bind/src/json.rs
@@ -141,7 +141,7 @@ fn cell_to_json(col: &dyn Array, row: usize) -> serde_json::Value {
                 .as_any()
                 .downcast_ref::<array::TimestampSecondArray>()
                 .expect("Arrow type mismatch: expected TimestampSecondArray");
-            let ts = arr.value(row) as i64;
+            let ts = arr.value(row);
             let dt = chrono::DateTime::from_timestamp(ts, 0);
             match dt {
                 Some(d) => serde_json::Value::String(d.to_rfc3339()),
